% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomGradient.R
\name{RandomGradient}
\alias{RandomGradient}
\title{Create a random gradient}
\usage{
RandomGradient(n_col, n_row, plot = TRUE)
}
\arguments{
\item{n_col}{integer. Number of columns in the image.}

\item{n_row}{integer. Number of rows in the image.}

\item{plot}{logical. Whether the function plot the image. Default is TRUE.}
}
\value{
A pixel image (object of class "im")
}
\description{
Create a random gradient
}
\examples{
RandomGradient(20, 20)

}
